#VMS-3422 it is allowed now for http client to generate cloud nonce itself
Internal: Compiling fixed for Linux.
VMS-3441 udt: android pthread_cond monotonic fix
fixed compile error

<==== Mercurial data ====>
Branch: dev_3.0.0_gui

+++ mediaserver_core/src/rtsp/rtsp_connection.cpp
@@ -429,7 +429,7 @@ void QnRtspConnectionProcessor::initResponse(int code, const QString& message)
 void QnRtspConnectionProcessor::generateSessionId()
 {
     Q_D(QnRtspConnectionProcessor);
-    d->sessionId = QString::number(reinterpret_cast<quint32>(d->socket.data()));
+    d->sessionId = QString::number(reinterpret_cast<uintptr_t>(d->socket.data()));
     d->sessionId += QString::number(nx::utils::random::number(0));
 }
